small change to the return dtype of unravel_index of torch

+++ ivy/functional/backends/torch/experimental/statistical.py
@@ -144,7 +144,7 @@ def unravel_index(
     *,
     out: Optional[torch.Tensor] = None,
 ) -> Tuple:
-    temp = indices.to(torch.int64)
+    temp = indices.to(torch.int32)
     output = []
     for dim in reversed(shape):
         output.append(temp % dim)
